refactor(discriminators): rename and move discriminator files to separate folder
Change recompose[...Resolver] to prepare[...Resolver]
